requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
python-jobspy>=1.1.80
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.9.0
playwright>=1.40.0
